Squashed commit of the following:
commit c45470e6b315fc692b776c9bdc58eb7254fb66ea
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Thu May 11 09:03:57 2023 +0200

    Basic user class complete

    User class has CRUD

commit d8db3159808ab7d689673438723308c903a243dc
Author: fewpixels <1038212@mborijnland.nl>
Date:   Thu May 11 09:35:24 2023 +0200

    small update to code

    added one or two methods for dashboard

commit 7912d56a7a8c7ca107e458ec53299f86c3707f13
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Wed May 10 13:46:33 2023 +0200

    Merge dashboard with main application

    very, very basic implementation of a "scans" dashboard and changed code to align the structure with the rest

commit ec10be3c759f1bfe05fd9ae049558fdca3a888de
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Sat Apr 29 11:35:33 2023 +0200

    added options for the scanner

    this could be potentially used to reduce the stress on the cpu

commit 11b5eefb17e99676686ae0a1c8f2ffe98cddde0b
Author: fewpixels <1038212@mborijnland.nl>
Date:   Mon Apr 24 16:35:17 2023 +0200

    undoing temp commit

    final test

commit f20c435d42cb70edb5c88f6db05e66058aedf2e7
Author: fewpixels <1038212@mborijnland.nl>
Date:   Mon Apr 24 16:34:51 2023 +0200

    temporary test commit

commit 3f27fe319c895465a2a076d49cefb360ad2f647b
Author: fewpixels <1038212@mborijnland.nl>
Date:   Mon Apr 24 16:29:26 2023 +0200

    minor changes to make it accessible for linux

    changed spelling of certain lines for linux systems running lampp to give access to

commit 9c3cc756f4a80a5fdc459152a7703a3ff7f81b25
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Fri Apr 21 13:20:19 2023 +0200

    adjusted data type for total hours

    adjusted a datatype in order to increase precision when registering hours

commit c1aa5238726573d50a608ba703331f094cacb842
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Fri Apr 21 09:29:43 2023 +0200

    add folder generated

    added a file to create the folder generated

commit 83b51aaaad0672c5e75a7165a6c28dce69929cf7
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Thu Apr 20 14:23:51 2023 +0200

    php qr generator implemented into application

    users are now able to add new users and generate QR codes for the new users from within the application

commit 081a65e3b8a78e044b9614e70235692d5f9c3243
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Tue Apr 18 10:35:20 2023 +0200

    PHP qr generator library

    added QR generator.

commit 7f3d1ce2fc7d540792fd22687c7b8ea4efd95c4b
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Tue Apr 18 09:36:34 2023 +0200

    Moved style and image folder

    to make the application whole

commit 5668f58ecef07f096beb8b4bf820e7dfb39b40b9
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Wed Apr 12 13:19:54 2023 +0200

    clean up data lookup file

commit b616e11c56dd092ccbb942e3659ae74a37e43968
Merge: 8abd050 2b9f929
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Wed Apr 12 09:35:44 2023 +0200

    Merge branch 'continuousCheckIn' into ProgressLogin

commit 2b9f92934ad7b23fa55494afba59ad51568311bb
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Tue Apr 4 14:38:38 2023 +0200

    changed logics

    user can now check in and check out throughout the day

    also changed the way the code picks up the records from the database

    changed forgot hours from 8 to 1

    changed some logics in the checks throughout the code

commit 3cf7a0212b811f40a1e2a0713bd08211e72f336c
Author: fewpixels <74591245+fewpixels@users.noreply.github.com>
Date:   Tue Apr 4 13:15:58 2023 +0200

    removal of unnecesary files

+++ CreateDatabase.sql
@@ -14,7 +14,7 @@ CREATE TABLE scan(
     userID INT(10) NOT NULL,
     inlogTijd DATETIME,
     uitlogTijd DATETIME,
-    totaalTijd INT
+    totaalTijd DECIMAL(5,2)
 );
 
 ALTER TABLE scan
